Asset, finishing

+++ app/Http/Controllers/DepreciationMethodController.php
@@ -2,23 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\AssetType;
 use Illuminate\Http\Request;
 use App\Models\Depreciation;
 use App\Models\Company;
 use App\Models\Department;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class DepreciationMethodController extends Controller
 {
     public function index()
     {
         $depreciations = Depreciation::orderBy('depreciation_code', 'asc')->get();
-        $companies = Company::orderBy('id', 'asc')->get();
-        $departments = Department::orderBy('id', 'asc')->get();
         return view('master.depreciation', [
             'depreciations' => $depreciations,
-            'companies' => $companies,
-            'departments' => $departments,
         ]);
     }
 
@@ -28,16 +26,15 @@ class DepreciationMethodController extends Controller
         $request->validate([
             'depreciation_code' => 'required',
             'depreciation_name' => 'required',
-            'company_code' => 'required',
-            'department_code' => 'required',
         ]);
 
+        $company =Company::first();
         if (Depreciation::where('depreciation_code', $request->depreciation_code)->count() < 1) {
             Depreciation::create([
                 'depreciation_code' => $request->depreciation_code,
                 'depreciation_name' => $request->depreciation_name,
-                'company_code' => $request->company_code,
-                'department_code' => $request->department_code,
+                'company_code' => $company->company_code,
+                'department_code' => 'DP01',
                 'created_by' => Auth::user()->username,
                 'updated_by' => Auth::user()->username,
             ]);
@@ -49,20 +46,9 @@ class DepreciationMethodController extends Controller
 
     public function update(Request $request, $id)
     {
-        // Uncomment the validation if needed
-        $request->validate([
-            'depreciation_code' => 'required',
-            'depreciation_name' => 'required',
-            'company_code' => 'required',
-            'department_code' => 'required',
-        ]);
 
         $depreciation = Depreciation::where('depreciation_code', $id)->update([
-            'depreciation_name' => $request->depreciation_code,
             'depreciation_name' => $request->depreciation_name,
-            'company_code' => $request->company_code,
-            'department_code' => $request->department_code,
-            'created_by' => Auth::user()->username,
             'updated_by' => Auth::user()->username,
         ]);
 
@@ -71,6 +57,10 @@ class DepreciationMethodController extends Controller
 
     public function delete($id)
     {
+        $asset = AssetType::where('depreciation_code',$id)->exists();
+        if($asset){
+            return redirect()->back()->with('error', 'Sudah digunakan pada jenis asset!');
+        }
         $depreciation = Depreciation::where('depreciation_code',$id);
         $depreciation->delete();
         return redirect()->back()->with('success', 'Depreciation Method deleted successfully!');
